+++ Jogo/src/main/java/br/ufscar/dc/cg/jogo/RGBColor.java
@@ -2,9 +2,9 @@ package br.ufscar.dc.cg.jogo;
 
 class RGBColor {
 
-    float R, G, B;
+    double R, G, B;
 
-    RGBColor(float r, float g, float b) {
+    RGBColor(double r, double g, double b) {
         R = r;
         G = g;
         B = b;
